New class S3Mapping that will serve the S3 objects
